Fix recall buttons crash: cast SM-2 interval to INTEGER, add error handling
The updateSm2 query used `CURRENT_DATE + $2` which PostgreSQL couldn't
resolve (ambiguous operator for date + unknown). Added explicit
`$2::INTEGER` cast. Also added try/catch to recall handler and a global
bot.catch() to prevent any unhandled callback errors from crashing the
process (caused 9+ pod restarts).

+++ src/db/progress.ts
@@ -314,7 +314,7 @@ export async function updateSm2(chatId: number, wordValue: string, quality: numb
 
   await pool.query(
     `UPDATE sent_words SET ease_factor = $1, interval_days = $2, repetitions = $3,
-     next_review = CURRENT_DATE + $2
+     next_review = CURRENT_DATE + $2::INTEGER
      WHERE chat_id = $4 AND word_value = $5`,
     [ef, interval, reps, chatId, wordValue],
   );
